+++ app/streaming/video/ffmpeg-renderers/dxva2.cpp
@@ -672,6 +672,11 @@ void DXVA2Renderer::notifyOverlayUpdated(Overlay::OverlayType type)
         renderRect.x = 0;
         renderRect.y = m_DisplayHeight - newSurface->h;
     }
+    else if (type == Overlay::OverlaySessionTimer) {
+        // Top center
+        renderRect.x = (m_DisplayWidth - newSurface->w) / 2.0f;
+        renderRect.y = 0;
+    }
     else if (type == Overlay::OverlayDebug) {
         // Top left
         renderRect.x = 0;