+++ app/streaming/video/ffmpeg-renderers/sdlvid.cpp
@@ -240,6 +240,13 @@ void SdlRenderer::renderOverlay(Overlay::OverlayType type)
                 m_OverlayRects[type].x = 0;
                 m_OverlayRects[type].y = viewportRect.h - newSurface->h;
             }
+            else if (type == Overlay::OverlaySessionTimer) {
+                // Top center
+                SDL_Rect viewportRect;
+                SDL_RenderGetViewport(m_Renderer, &viewportRect);
+                m_OverlayRects[type].x = (viewportRect.w - newSurface->w) / 2;
+                m_OverlayRects[type].y = 0;
+            }
             else if (type == Overlay::OverlayDebug) {
                 // Top left
                 m_OverlayRects[type].x = 0;