+++ app/streaming/video/ffmpeg-renderers/vt_avsamplelayer.mm
@@ -413,6 +413,9 @@ public:
             case Overlay::OverlayDebug:
                 [m_OverlayTextFields[type] setAlignment:NSTextAlignmentLeft];
                 break;
+            case Overlay::OverlaySessionTimer:
+                [m_OverlayTextFields[type] setAlignment:NSTextAlignmentCenter];
+                break;
             case Overlay::OverlayStatusUpdate:
                 [m_OverlayTextFields[type] setAlignment:NSTextAlignmentRight];
                 break;